+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 This release includes bug fixes but also some new cool features.
 
 ### :star2: News
-
+* Entities are now accessed on the `Message` object in Intents instead of 
+  being accessed on the Intent itself. 
+  Accessing Entities on the Intent is supported until 1.2.0 and will raise 
+  a deprecation warning.
+* 
 
 ### **🐛 Splatted bugs and corrected issues** 
 * Fixes [#46](https://github.com/dotchetter/Pyttman/issues/46)